+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/colorcontrol/ColorLoopSetCommand.java
@@ -214,11 +214,11 @@ public class ColorLoopSetCommand extends ZclColorControlCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        updateFlags = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        action = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        direction = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        transitionTime = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        startHue = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        updateFlags = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        action = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        direction = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        transitionTime = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        startHue = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
