+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/identify/IdentifyQueryResponse.java
@@ -98,7 +98,7 @@ public class IdentifyQueryResponse extends ZclIdentifyCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        identifyTime = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        identifyTime = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
